backend/packages/harness/pyproject.toml
[project]
name = "deerflow-harness"
version = "0.1.0"
description = "DeerFlow agent harness framework"
requires-python = ">=3.12"
dependencies = [
    "agent-client-protocol>=0.4.0",
    "agent-sandbox>=0.0.19",
    "dotenv>=0.9.9",
    "httpx>=0.28.0",
    "kubernetes>=30.0.0",
    "langchain>=1.2.3",
    "langchain-anthropic>=1.3.4",
    "langchain-deepseek>=1.0.1",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=1.1.7",
    "langfuse>=3.4.1",
    "langgraph>=1.0.6,<1.0.10",
    "langgraph-api>=0.7.0,<0.8.0",
    "langgraph-cli>=0.4.14",
    "langgraph-runtime-inmem>=0.22.1",
    "markdownify>=1.2.2",
    "markitdown[all,xlsx]>=0.0.1a2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "readabilipy>=0.3.0",
    # "tavily-python>=0.7.17",  # Optional: Tavily API
    # "firecrawl-py>=1.15.0",  # Optional: Firecrawl API
    "tiktoken>=0.8.0",
    "ddgs>=9.10.0",
    "duckdb>=1.4.4",
    "langchain-google-genai>=4.2.1",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-sdk>=0.1.51",
    # 智能家居额外依赖
    "paho-mqtt>=2.0.0",  # MQTT 客户端
    # "HAP-python>=3.0.0",  # HomeKit 支持（可选）
    "requests>=2.31.0",  # HTTP 请求
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["deerflow"]
